Working on some error handling, less chatty for now.
Also yields actions up for connection problems so we can do things. This
also fixes an issue I noticed with the pendingExecution promise we use
to serialize device calls. One is updating the promise to the last and
also being sure to ignore errors from older requests when signing up a
new one. Things seem to work better now. TOOD: Not sure if this is
just going to grow and grow, though. Maybe check for a "done" promise?